# docs/reference/openapi-v1.json
{
  "openapi": "3.1.1",
...
  "info": {
    "title": "THE CAPTION 資産入力 API",
    "version": "0.1.0",
    "description": "段階Bの部分実装。Market Units GET/PUT、input-schema、healthを実装。External Assets/Portfolio Basisは契約案でv1未実装。汎用取込・入力元管理・Monex接続は後続段階。HTTPとデータの共通規則・UI対応はapi-v1.mdを参照。"
  },
  "servers": [
    {
      "url": "/api/v1",
      "description": "同一origin。開発用run.sh collection-web-devの既定はhttp://127.0.0.1:3101。"
    }
  ],
  "security": [
    {
      "serviceToken": []
    },
    {
      "browserSession": []
    }
  ],
...
  "paths": {
    "/market-units": {
      "get": {
        "tags": [
          "Market Units"
        ],
        "operationId": "getMarketUnits",
        "summary": "Market Unitsの全入力を取得",
        "description": "市場連動資産の登録済み全行を、一貫したrevisionのitems配列として取得する。保存順、7編集項目、不変asset_id、互換asset_keyを返す。IDはnullにならない。ページ分割しない。",
        "x-required-permissions": [
          "market-units:read"
        ],
        "responses": {
          "200": {
            "description": "一貫した入力document。ETagはこのdocumentの強いvalidator。",
            "headers": {
              "Cache-Control": {
                "$ref": "#/components/headers/NoStore"
              },
              "ETag": {
                "$ref": "#/components/headers/ETag"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MarketUnitsDocument"
                },
                "example": {
                  "revision": "rev_mu_3",
                  "storage_state": "ready",
                  "updated_at": "2026-09-08T18:00:00+09:00",
                  "items": [
                    {
                      "asset_id": "00000000-0000-4000-8000-000000000001",
                      "name": "国内株サンプル",
                      "asset_class": "JP_STOCK",
                      "currency": "JPY",
                      "units": "123",
                      "source_symbol": "EXAMPLE",
                      "audit_match_key": "sample-equity",
                      "csv_url": "",
                      "asset_key": "sample-equity"
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "429": {
            "$ref": "#/components/responses/TooManyRequests"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          },
          "503": {
            "$ref": "#/components/responses/ServiceUnavailable"
          }
        },
        "x-implementation-status": "implemented"
      },
      "put": {
        "tags": [
          "Market Units"
        ],
        "operationId": "replaceMarketUnits",
        "summary": "Market Unitsの全入力を条件付きで置換",
        "description": "配列順を保持する。全既存asset_idは同じresourceに存在し、一意であること。省略した既存IDは削除、nullは新規。互換asset_keyも正規化後に一意。enabled等の非UI既存metadataは保持する。 前後内容と履歴・再送結果は一緒に永続化する。失敗時に一部だけ保存しない。成功後はGETで正規化済み内容と次回ETagを取得。",
        "x-required-permissions": [
          "market-units:read",
          "market-units:replace"
        ],
        "x-conditional-permissions": [
          {
            "when": "clear_all=true",
            "allOf": [
              "market-units:clear"
            ]
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/IfMatch"
          },
          {
            "$ref": "#/components/parameters/IdempotencyKey"
          },
          {
            "$ref": "#/components/parameters/CsrfToken"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReplaceMarketUnits"
              },
              "examples": {
                "newEntry": {
                  "summary": "架空の新規入力。既存行はIDを維持して併記する。",
                  "value": {
                    "items": [
                      {
                        "asset_id": null,
                        "name": "国内株サンプル",
                        "asset_class": "JP_STOCK",
                        "currency": "JPY",
                        "units": "123",
                        "source_symbol": "EXAMPLE",
                        "audit_match_key": "sample-equity",
                        "csv_url": ""
                      }
                    ],
                    "clear_all": false
                  }
                },
                "clearAll": {
                  "summary": "明示的な全消去。別clear権限が必要。",
                  "value": {
                    "items": [],
                    "clear_all": true
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "保存結果または同一要求の再送結果。新IDも含めた全documentを返す。ETag/Last-Modifiedは付けない。resource.revisionはIf-Matchに転用しない。",
            "headers": {
              "Cache-Control": {
                "$ref": "#/components/headers/NoStore"
              },
              "Idempotency-Replayed": {
                "$ref": "#/components/headers/Replay"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MarketUnitsReplacementResult"
                },
                "example": {
                  "changed": true,
                  "change_id": "00000000-0000-4000-8000-000000000050",
                  "resource": {
                    "revision": "rev_mu_3",
                    "storage_state": "ready",
                    "updated_at": "2026-09-08T18:00:00+09:00",
                    "items": [
                      {
                        "asset_id": "00000000-0000-4000-8000-000000000001",
                        "name": "国内株サンプル",
                        "asset_class": "JP_STOCK",
                        "currency": "JPY",
                        "units": "123",
                        "source_symbol": "EXAMPLE",
                        "audit_match_key": "sample-equity",
                        "csv_url": "",
                        "asset_key": "sample-equity"
                      }
                    ]
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "412": {
            "$ref": "#/components/responses/PreconditionFailed"
          },
          "413": {
            "$ref": "#/components/responses/PayloadTooLarge"
          },
          "415": {
            "$ref": "#/components/responses/UnsupportedMediaType"
          },
          "422": {
            "$ref": "#/components/responses/UnprocessableContent"
          },
          "428": {
            "$ref": "#/components/responses/PreconditionRequired"
          },
          "429": {
            "$ref": "#/components/responses/TooManyRequests"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          },
          "503": {
            "$ref": "#/components/responses/ServiceUnavailable"
          }
        },
        "x-implementation-status": "implemented"
      }
    },
    "/external-assets": {
      "get": {
        "tags": [
          "External Assets"
        ],
        "operationId": "getExternalAssets",
        "summary": "External Assetsの全入力を取得",
        "description": "外部資産の全月とdefaultを、一貫したrevisionのmonths mapとして取得する。各月の配列順と不変entry_idを返す。明示的な空月を保持する。ページ分割しない。",
        "x-required-permissions": [
          "external-assets:read"
        ],
        "responses": {
          "200": {
            "description": "一貫した入力document。ETagはこのdocumentの強いvalidator。",
            "headers": {
              "Cache-Control": {
                "$ref": "#/components/headers/NoStore"
              },
              "ETag": {
                "$ref": "#/components/headers/ETag"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExternalAssetsDocument"
                },
                "example": {
                  "revision": "rev_ea_3",
                  "storage_state": "ready",
                  "updated_at": "2026-09-08T18:00:00+09:00",
                  "months": {
                    "2026-09": {
                      "items": [
                        {
                          "entry_id": "00000000-0000-4000-8000-000000000001",
                          "category": "CASH_EXTERNAL",
                          "amount": "1000000",
                          "name": "外部資産サンプル"
                        }
                      ]
                    },
                    "default": {
                      "items": []
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "429": {
            "$ref": "#/components/responses/TooManyRequests"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          },
          "503": {
            "$ref": "#/components/responses/ServiceUnavailable"
          }
        },
        "x-implementation-status": "proposed"
      },
      "put": {
        "tags": [
          "External Assets"
        ],
        "operationId": "replaceExternalAssets",
        "summary": "External Assetsの全入力を条件付きで置換",
        "description": "months map全体を置換。entry_idは全月で一意、移動時も保持。空の月items=[]はdefaultを抑止する明示入力として保持する。月key省略は削除。省略した既存IDは削除。各月の配列順を保持。全月総数5000以下。JPY絶対額のみ。 前後内容と履歴・再送結果は一緒に永続化する。失敗時に一部だけ保存しない。成功後はGETで正規化済み内容と次回ETagを取得。",
        "x-required-permissions": [
          "external-assets:read",
          "external-assets:replace"
        ],
        "x-conditional-permissions": [
          {
            "when": "clear_all=true",
            "allOf": [
              "external-assets:clear"
            ]
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/IfMatch"
          },
          {
            "$ref": "#/components/parameters/IdempotencyKey"
          },
          {
            "$ref": "#/components/parameters/CsrfToken"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReplaceExternalAssets"
              },
              "examples": {
                "newEntry": {
                  "summary": "架空の新規入力。既存行はIDを維持して併記する。",
                  "value": {
                    "months": {
                      "2026-09": {
                        "items": [
                          {
                            "entry_id": null,
                            "category": "CASH_EXTERNAL",
                            "amount": "1000000",
                            "name": "外部資産サンプル"
                          }
                        ]
                      },
                      "default": {
                        "items": []
                      }
                    },
                    "clear_all": false
                  }
                },
                "clearAll": {
                  "summary": "明示的な全消去。別clear権限が必要。",
                  "value": {
                    "months": {},
                    "clear_all": true
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "保存結果または同一要求の再送結果。新IDも含めた全documentを返す。ETag/Last-Modifiedは付けない。resource.revisionはIf-Matchに転用しない。",
            "headers": {
              "Cache-Control": {
                "$ref": "#/components/headers/NoStore"
              },
              "Idempotency-Replayed": {
                "$ref": "#/components/headers/Replay"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExternalAssetsReplacementResult"
                },
                "example": {
                  "changed": true,
                  "change_id": "00000000-0000-4000-8000-000000000050",
                  "resource": {
                    "revision": "rev_ea_3",
                    "storage_state": "ready",
                    "updated_at": "2026-09-08T18:00:00+09:00",
                    "months": {
                      "2026-09": {
                        "items": [
                          {
                            "entry_id": "00000000-0000-4000-8000-000000000001",
                            "category": "CASH_EXTERNAL",
                            "amount": "1000000",
                            "name": "外部資産サンプル"
                          }
                        ]
                      },
                      "default": {
                        "items": []
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "412": {
            "$ref": "#/components/responses/PreconditionFailed"
          },
          "413": {
            "$ref": "#/components/responses/PayloadTooLarge"
          },
          "415": {
            "$ref": "#/components/responses/UnsupportedMediaType"
          },
          "422": {
            "$ref": "#/components/responses/UnprocessableContent"
          },
          "428": {
            "$ref": "#/components/responses/PreconditionRequired"
          },
          "429": {
            "$ref": "#/components/responses/TooManyRequests"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          },
          "503": {
            "$ref": "#/components/responses/ServiceUnavailable"
          }
        },
        "x-implementation-status": "proposed"
      }
    },
    "/portfolio-basis": {
      "get": {
        "tags": [
          "Portfolio Basis"
        ],
        "operationId": "getPortfolioBasis",
        "summary": "Portfolio Basisの全入力を取得",
        "description": "総取得原価の全月とdefaultを、一貫したrevisionのmonths mapとして取得する。月ごとの正の金額と不変entry_idを返す。ページ分割しない。",
        "x-required-permissions": [
          "portfolio-basis:read"
        ],
        "responses": {
          "200": {
            "description": "一貫した入力document。ETagはこのdocumentの強いvalidator。",
            "headers": {
              "Cache-Control": {
                "$ref": "#/components/headers/NoStore"
              },
              "ETag": {
                "$ref": "#/components/headers/ETag"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PortfolioBasisDocument"
                },
                "example": {
                  "revision": "rev_pb_3",
                  "storage_state": "ready",
                  "updated_at": "2026-09-08T18:00:00+09:00",
                  "months": {
                    "2026-09": {
                      "entry_id": "00000000-0000-4000-8000-000000000002",
                      "total_acquisition_cost_jpy": "1500000"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "429": {
            "$ref": "#/components/responses/TooManyRequests"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          },
          "503": {
            "$ref": "#/components/responses/ServiceUnavailable"
          }
        },
        "x-implementation-status": "proposed"
      },
      "put": {
        "tags": [
          "Portfolio Basis"
        ],
        "operationId": "replacePortfolioBasis",
        "summary": "Portfolio Basisの全入力を条件付きで置換",
        "description": "months map全体を置換。entry_idは月を跨いで一意。月移動では元entry_idを維持し移動元keyを省略。既存移動先を置換する場合、そのIDはrequestから省略されるため明示的に削除される。未指定月やdefaultの意味を変えない。 前後内容と履歴・再送結果は一緒に永続化する。失敗時に一部だけ保存しない。成功後はGETで正規化済み内容と次回ETagを取得。",
        "x-required-permissions": [
          "portfolio-basis:read",
          "portfolio-basis:replace"
        ],
        "x-conditional-permissions": [
          {
            "when": "clear_all=true",
            "allOf": [
              "portfolio-basis:clear"
            ]
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/IfMatch"
          },
          {
            "$ref": "#/components/parameters/IdempotencyKey"
          },
          {
            "$ref": "#/components/parameters/CsrfToken"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReplacePortfolioBasis"
              },
              "examples": {
                "newEntry": {
                  "summary": "架空の新規入力。既存行はIDを維持して併記する。",
                  "value": {
                    "months": {
                      "2026-09": {
                        "entry_id": null,
                        "total_acquisition_cost_jpy": "1500000"
                      }
                    },
                    "clear_all": false
                  }
                },
                "clearAll": {
                  "summary": "明示的な全消去。別clear権限が必要。",
                  "value": {
                    "months": {},
                    "clear_all": true
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "保存結果または同一要求の再送結果。新IDも含めた全documentを返す。ETag/Last-Modifiedは付けない。resource.revisionはIf-Matchに転用しない。",
            "headers": {
              "Cache-Control": {
                "$ref": "#/components/headers/NoStore"
              },
              "Idempotency-Replayed": {
                "$ref": "#/components/headers/Replay"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PortfolioBasisReplacementResult"
                },
                "example": {
                  "changed": true,
                  "change_id": "00000000-0000-4000-8000-000000000050",
                  "resource": {
                    "revision": "rev_pb_3",
                    "storage_state": "ready",
                    "updated_at": "2026-09-08T18:00:00+09:00",
                    "months": {
                      "2026-09": {
                        "entry_id": "00000000-0000-4000-8000-000000000002",
                        "total_acquisition_cost_jpy": "1500000"
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "412": {
            "$ref": "#/components/responses/PreconditionFailed"
          },
          "413": {
            "$ref": "#/components/responses/PayloadTooLarge"
          },
          "415": {
            "$ref": "#/components/responses/UnsupportedMediaType"
          },
          "422": {
            "$ref": "#/components/responses/UnprocessableContent"
          },
          "428": {
            "$ref": "#/components/responses/PreconditionRequired"
          },
          "429": {
            "$ref": "#/components/responses/TooManyRequests"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          },
          "503": {
            "$ref": "#/components/responses/ServiceUnavailable"
          }
        },
        "x-implementation-status": "proposed"
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "serviceToken": {
        "type": "http",
        "scheme": "bearer",
        "description": "THE CAPTION専用の不透明token。JWT形式・OAuth flowを仮定しない。token発行/失効とsession作成はPhase Bの認証設計で別途固定。Monex用tokenは受理しない。"
      },
      "browserSession": {
        "type": "apiKey",
        "in": "cookie",
        "name": "caption_session",
        "description": "認証済みWebUI session。HttpOnly、HTTPS環境ではSecure、SameSite=Strict。PUTはsessionにbindされたX-CSRF-Tokenを検証。"
      }
    },
...
    "headers": {
      "ETag": {
        "description": "このリソース表現の強いvalidator。必ず引用符を含む。",
        "schema": {
          "type": "string",
          "pattern": "^\"[A-Za-z0-9:_-]+\"$"
        }
      },
      "NoStore": {
        "description": "資産情報・session依存情報・エラーをキャッシュしない。",
        "schema": {
          "type": "string",
          "const": "no-store"
        }
      },
...
    },
    "responses": {
      "BadRequest": {
        "description": "不正JSON（重複object keyを含む）、不正header、認証方式の同時指定。",
        "headers": {
          "Cache-Control": {
            "$ref": "#/components/headers/NoStore"
          }
        },
        "content": {
          "application/problem+json": {
            "schema": {
              "allOf": [
                {
                  "$ref": "#/components/schemas/Problem"
                },
                {
                  "type": "object",
                  "properties": {
                    "status": {
                      "const": 400
                    }
                  }
                }
              ]
            },
            "example": {
              "type": "about:blank",
              "title": "Bad Request",
              "status": 400,
              "detail": "不正JSON（重複object keyを含む）、不正header、認証方式の同時指定。",
              "instance": "urn:uuid:00000000-0000-4000-8000-000000000099",
              "code": "invalid_request"
            }
          }
        }
      },
      "Unauthorized": {
        "description": "認証なし・期限切れ。",
        "headers": {
          "Cache-Control": {
            "$ref": "#/components/headers/NoStore"
          },
          "WWW-Authenticate": {
            "$ref": "#/components/headers/WWWAuthenticate"
          }
        },
        "content": {
          "application/problem+json": {
            "schema": {
              "allOf": [
                {
                  "$ref": "#/components/schemas/Problem"
                },
                {
                  "type": "object",
                  "properties": {
                    "status": {
                      "const": 401
                    }
                  }
                }
              ]
            },
            "example": {
              "type": "about:blank",
              "title": "Unauthorized",
              "status": 401,
              "detail": "認証なし・期限切れ。",
              "instance": "urn:uuid:00000000-0000-4000-8000-000000000099",
              "code": "authentication_required"
            }
          }
        }
      },
      "Forbidden": {
        "description": "権限不足、sessionのCSRF検証失敗。",
        "headers": {
          "Cache-Control": {
            "$ref": "#/components/headers/NoStore"
          }
        },
        "content": {
          "application/problem+json": {
            "schema": {
              "allOf": [
                {
                  "$ref": "#/components/schemas/Problem"
                },
                {
                  "type": "object",
                  "properties": {
                    "status": {
                      "const": 403
                    }
                  }
                }
              ]
            },
            "example": {
              "type": "about:blank",
              "title": "Forbidden",
              "status": 403,
              "detail": "権限不足、sessionのCSRF検証失敗。",
              "instance": "urn:uuid:00000000-0000-4000-8000-000000000099",
              "code": "permission_denied"
            }
          }
        }
      },
...
      "TooManyRequests": {
        "description": "rate limit。Retry-After後に同一requestを再送可能。",
        "headers": {
          "Cache-Control": {
            "$ref": "#/components/headers/NoStore"
          },
          "Retry-After": {
            "$ref": "#/components/headers/RetryAfter"
          }
        },
        "content": {
          "application/problem+json": {
            "schema": {
              "allOf": [
                {
                  "$ref": "#/components/schemas/Problem"
                },
                {
                  "type": "object",
                  "properties": {
                    "status": {
                      "const": 429
                    }
                  }
                }
              ]
            },
            "example": {
              "type": "about:blank",
              "title": "Too Many Requests",
              "status": 429,
              "detail": "rate limit。Retry-After後に同一requestを再送可能。",
              "instance": "urn:uuid:00000000-0000-4000-8000-000000000099",
              "code": "rate_limited"
            }
          }
        }
      },
...
    },
    "schemas": {
...
      "MarketUnitsDocument": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "revision",
          "storage_state",
          "updated_at",
          "items"
        ],
        "properties": {
          "revision": {
            "$ref": "#/components/schemas/Revision"
          },
          "storage_state": {
            "type": "string",
            "enum": [
              "ready",
              "uninitialized"
            ]
          },
          "updated_at": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time",
            "description": "当該revisionの保存時刻。未初期化だけnull。GET時刻ではない。"
          },
          "items": {
            "type": "array",
            "maxItems": 5000,
            "items": {
              "$ref": "#/components/schemas/MarketUnit"
            }
          }
        }
      },
...
    }
  },
...
}